ner-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6

transformers==4.48.3
torch==2.5.1

python-ulid==3.0.0
httpx>=0.27.0

protobuf>=4.25.0
tiktoken>=0.7.0
sentencepiece>=0.2.0
